Rework Page class

+++ index.php
@@ -137,39 +137,27 @@ function list_pages($dir = "/") {
             $list[] = pagename($entry);
     }
 
-    return count($list)>0 ? $list : false;
+    return count($list)>0 ? $list : null;
 }
 
 class Page {
-    public $version;
-    public $text;
-    public $time;
-    public $header;
-    
-    public function fetch($_ = null, $v = null) {
-        $page = new self;
-
-        if ($v != null)
-            $v = filename($_, "pages/v/")."~".$v;
-        elseif (is_link(filename($_)))
-            $v = readlink(filename($_));
+    public function fetch($_, $v = null) {
+        if     ($v !== null)           $v = filename($_, "pages/v/")."~".$v;
+        elseif (is_link(filename($_))) $v = readlink(filename($_));
 
         if (file_exists($v)) {
+            $page = new self;
             $page->version = explode("~", $v)[1];
             $page->text = file_get_contents($v);
             $page->time = filemtime($v);
 
-            # get the header
-            if (substr($page->text, 0, 2) == "\0:") {
+            if (substr($page->text, 0, 2) == "\0:") { # get the header, if it exists
                 $t = explode("\n", $page->text);
                 parse_str(substr(array_shift($t), 2), $page->header);
                 $page->text = join("\n", $t);
             }
-
             return $page;
         }
-
-        return false;
     }
 
     public function store($name, $contents, $header) {
@@ -177,23 +165,20 @@ class Page {
 
         if (file_exists($link)) unlink($link);
         
-        if (is_link($link)) {
-            $last = readlink($link);
+        if (is_link($link) && $last = readlink($link)) {
             preg_match("/~\d+/", $last, $next);
             $next = filename($name, "pages/v/")
-                  . "~" . (substr(reset($next), 1) + 1);
+                  ."~". (substr(reset($next), 1) + 1);
             unlink($link);
         }
 
         if (empty($next)) 
-            $next = filename($name, "pages/v/") . "~0";
+            $next = filename($name, "pages/v/") ."~0";
 
-        $contents = "\0:" . http_build_query($header) . "\n" . $contents;
+        $contents = "\0:". http_build_query($header) ."\n". $contents;
 
         if (@file_put_contents($next, $contents))
             return symlink($next, $link);
-
-        return false;
     }
 }
